pyproject.toml
[tool.poetry]
name = "bookshop"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = {extras = ["email"], version = "^2.6.3"}
fastapi = "^0.110.0"
icecream = "^2.1.3"
asyncpg = "^0.29.0"
orjson = "^3.9.15"
sqlalchemy = "^2.0.27"
pydantic-settings = "^2.2.1"
uvicorn = "^0.27.1"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
